# openApi/snackstasy-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Snackstasy API",
    "description": "API für das Essens-Bestellungs-Tool",
    "version": "1.0.0",
    "contact": {
      "name": "Snackstasy Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/v1",
      "description": "Local Development Server"
    }
  ],
  "tags": [
    {
      "name": "Login",
      "description": "Authentifizierung und Session-Management"
    },
    {
      "name": "Tickets",
      "description": "Ticket-Verwaltung"
    },
    {
      "name": "Balance",
      "description": "Guthaben-Verwaltung"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": ["Login"],
        "summary": "Login und Session erstellen",
        "description": "Authentifiziert einen Benutzer und erstellt eine Session",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login erfolgreich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falscher Benutzername",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Benutzer nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/login-check": {
      "post": {
        "tags": ["Login"],
        "summary": "Prüft ob ein Account existiert",
        "description": "Überprüft ob ein Account mit den angegebenen Daten existiert",
        "operationId": "loginCheck",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account existiert",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falscher Benutzername",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Benutzer nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": ["Login"],
        "summary": "Logout und Session beenden",
        "description": "Beendet die aktuelle Session des Benutzers",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Logout erfolgreich",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Keine aktive Session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/session": {
      "get": {
        "tags": ["Login"],
        "summary": "Session-Status prüfen",
        "description": "Gibt den aktuellen Session-Status zurück",
        "operationId": "getSession",
        "responses": {
          "200": {
            "description": "Session-Status",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/SessionResponse" },
                    { "$ref": "#/components/schemas/SessionNotLoggedIn" }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/tickets/all": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Alle Tickets abrufen",
        "description": "Gibt alle Tickets zurück",
        "operationId": "getAllTickets",
        "responses": {
          "200": {
            "description": "Liste aller Tickets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "500": {
            "description": "Interner Serverfehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/book": {
      "post": {
        "tags": ["Tickets"],
        "summary": "Neues Ticket buchen",
        "description": "Erstellt ein neues Ticket für einen Benutzer",
        "operationId": "bookTicket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket erfolgreich gebucht",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nicht autorisiert"
          }
        }
      }
    },
    "/balance/{ticket_id}": {
      "get": {
        "tags": ["Balance"],
        "summary": "Guthaben abrufen",
        "description": "Gibt das aktuelle Guthaben eines Tickets zurück",
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "description": "Die ID des Tickets",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aktuelles Guthaben",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "404": {
            "description": "Ticket nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/balance/{ticket_id}/update/{new_balance}": {
      "put": {
        "tags": ["Balance"],
        "summary": "Guthaben setzen",
        "description": "Setzt das Guthaben auf einen bestimmten Wert",
        "operationId": "updateBalance",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "description": "Die ID des Tickets",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "new_balance",
            "in": "path",
            "required": true,
            "description": "Das neue Guthaben",
            "schema": {
              "type": "number",
              "format": "decimal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guthaben erfolgreich aktualisiert",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "404": {
            "description": "Ticket nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/balance/{ticket_id}/add/{amount}": {
      "put": {
        "tags": ["Balance"],
        "summary": "Guthaben aufladen",
        "description": "Fügt einen Betrag zum Guthaben hinzu",
        "operationId": "addBalance",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "description": "Die ID des Tickets",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "path",
            "required": true,
            "description": "Der aufzuladende Betrag",
            "schema": {
              "type": "number",
              "format": "decimal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guthaben erfolgreich aufgeladen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "404": {
            "description": "Ticket nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/balance/{ticket_id}/remove/{amount}": {
      "put": {
        "tags": ["Balance"],
        "summary": "Guthaben abziehen",
        "description": "Zieht einen Betrag vom Guthaben ab",
        "operationId": "removeBalance",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "description": "Die ID des Tickets",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "path",
            "required": true,
            "description": "Der abzuziehende Betrag",
            "schema": {
              "type": "number",
              "format": "decimal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guthaben erfolgreich abgezogen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unzureichendes Guthaben",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "404": {
            "description": "Ticket nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["user_id", "username"],
        "properties": {
          "user_id": {
            "type": "string",
            "description": "Die Benutzer-ID (Ticket-ID)",
            "example": "12345"
          },
          "username": {
            "type": "string",
            "description": "Der Benutzername",
            "example": "max.mustermann"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "first_name": {
            "type": "string",
            "example": "Max"
          },
          "last_name": {
            "type": "string",
            "example": "Mustermann"
          },
          "balance": {
            "type": "number",
            "format": "float",
            "example": 25.50
          }
        }
      },
      "LoginCheckResponse": {
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean",
            "example": true
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "first_name": {
            "type": "string",
            "example": "Max"
          },
          "last_name": {
            "type": "string",
            "example": "Mustermann"
          },
          "balance": {
            "type": "number",
            "format": "float",
            "example": 25.50
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "logged_in": {
            "type": "boolean",
            "example": true
          },
          "user_id": {
            "type": "string",
            "example": "12345"
          },
          "username": {
            "type": "string",
            "example": "max.mustermann"
          },
          "first_name": {
            "type": "string",
            "example": "Max"
          },
          "last_name": {
            "type": "string",
            "example": "Mustermann"
          },
          "logged_in_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-09T14:30:00Z"
          }
        }
      },
      "SessionNotLoggedIn": {
        "type": "object",
        "properties": {
          "logged_in": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "BookRequest": {
        "type": "object",
        "required": ["first_name", "last_name"],
        "properties": {
          "user_id": {
            "type": "integer",
            "description": "Benutzer-ID (optional)"
          },
          "first_name": {
            "type": "string",
            "description": "Vorname",
            "example": "Max"
          },
          "last_name": {
            "type": "string",
            "description": "Nachname",
            "example": "Mustermann"
          },
          "birth_date": {
            "type": "string",
            "format": "date-time",
            "description": "Geburtsdatum"
          }
        }
      },
      "BookResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Die neu erstellte Ticket-ID",
            "example": 1
          },
          "first_name": {
            "type": "string",
            "example": "Max"
          },
          "last_name": {
            "type": "string",
            "example": "Mustermann"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "first_name": {
            "type": "string",
            "example": "Max"
          },
          "last_name": {
            "type": "string",
            "example": "Mustermann"
          },
          "balance": {
            "type": "number",
            "format": "float",
            "example": 25.50
          }
        }
      },
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "number",
            "format": "decimal",
            "example": 25.50
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation successful"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Error description"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      }
    }
  }
}